requirements-ib.txt
openai
fastapi
uvicorn
python-dotenv
pydantic
redis
guardrails
PyPDF2